pyproject.toml
[project]
name = "mcp-server-stdio-02"
version = "0.2.0"
description = "RAG MCP Server with Ollama embedding support"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu>=1.12.0",
    "fastmcp>=2.12.2",
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-text-splitters>=0.3.11",
    "pypdf>=6.0.0",
    "python-dotenv>=1.0.0",
    "langchain-ollama>=0.3.8",
    "unstructured>=0.8.0",
    "python-docx>=1.2.0",
    "pytest>=8.4.2",
    "langchain-openai>=0.3.35",
]
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
exclude = ["faiss_db", "faiss_db.*", "log", "log.*", "docs", "tests", "examples"]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
]
